+++ pkg/pipelines/debug.go
@@ -6,8 +6,7 @@ import (
 	"io"
 	"log/slog"
 	"os"
-	shell "workflow-engine/pkg/shell"
-	legacyShell "workflow-engine/pkg/shell/legacy"
+	"workflow-engine/pkg/shell"
 )
 
 type Debug struct {
@@ -39,21 +38,14 @@ func (d *Debug) Run() error {
 	// Collect errors for mandatory commands
 	commonOptions := []shell.OptionFunc{shell.WithDryRun(d.DryRunEnabled), shell.WithStdout(d.Stdout)}
 	errs := errors.Join(
-		legacyShell.GrypeCommand(nil, d.Stdout, d.Stderr).Version().WithDryRun(d.DryRunEnabled).Run(),
-		legacyShell.SyftCommand(nil, d.Stdout, d.Stderr).Version().WithDryRun(d.DryRunEnabled).Run(),
-		legacyShell.GitleaksCommand(nil, d.Stdout, d.Stderr).Version().WithDryRun(d.DryRunEnabled).Run(),
-		legacyShell.GatecheckCommand(nil, d.Stdout, d.Stderr).Version().WithDryRun(d.DryRunEnabled).Run(),
-		legacyShell.OrasCommand(nil, d.Stdout, d.Stderr).Version().WithDryRun(d.DryRunEnabled).Run(),
-		legacyShell.ClamScanCommand(nil, d.Stdout, d.Stderr).Version().WithDryRun(d.DryRunEnabled).Run(),
-		shell.GrypeVersion(commonOptions...).GetError("grype"),
-		shell.SyftVersion(commonOptions...).GetError("syft"),
-		shell.ClamScanVersion(commonOptions...).GetError("clamscan"),
-		shell.FreshClamVersion(commonOptions...).GetError("freshclam"),
+		shell.GrypeVersion(commonOptions...),
+		shell.SyftVersion(commonOptions...),
+		shell.FreshClamVersion(commonOptions...),
+		shell.ClamScanVersion(commonOptions...),
+		shell.GitLeaksVersion(commonOptions...),
+		shell.SemgrepVersion(commonOptions...),
+		shell.OrasVersion(commonOptions...),
 	)
-
-	// Just log errors for optional commands
-	legacyShell.PodmanCommand(nil, d.Stdout, d.Stderr).Version().WithDryRun(d.DryRunEnabled).RunLogErrorAsWarning()
-	legacyShell.DockerCommand(nil, d.Stdout, d.Stderr).Version().WithDryRun(d.DryRunEnabled).RunLogErrorAsWarning()
 
 	return errs
 }